+++ pages/library/collections.tsx
@@ -3,7 +3,7 @@ import Layout from '~/layout'
 
 const Collections: NextPage = () => {
   return (
-    <Layout heading="Home" content="Homepage of the legal engineering guild.">
+    <Layout heading="Home" content="Homepage of the legal engineering guild." criteria>
       <h1 className="text-3xl font-bold text-white underline">Collections inside Libarary</h1>
     </Layout>
   )
